backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
python-multipart
deep-translator
